+++ contracts/test/utils/math/ratio.ts
@@ -1,4 +1,4 @@
-import {RatioTest, RatioTest__factory} from '../../typechain';
+import {RatioTest, RatioTest__factory} from '../../../typechain';
 import {expect} from 'chai';
 import {ethers} from 'hardhat';
